move Promise and Promises to own folders
