+++ src/router/index.js
@@ -4,8 +4,8 @@ import HelloWorld from '@/components/HelloWorld'
 import Home from '@/components/Home'
 import Politician from '@/components/Politician/index'
 import Politicians from '@/components/Politicians'
-import Promise from '@/components/Promise'
-import Promises from '@/components/Promises'
+import Promise from '@/components/Promise/index'
+import Promises from '@/components/Promises/index'
 import About from '@/components/About'
 import SubmitPromise from '@/components/SubmitPromise'
 import Account from '@/components/Account'
